Create basic-component-mixins/src/lib folder for helper code, keeping the src folder pure for documented mixins. Moved microtask.js to lib folder.
